[stencils] Resolve macro conflict with "N"
When building with CUDA enabled, the weak/mmap.c would fail to build
with an obscure error message in the cuda.h header file.

In the stencils/stencils.h header file there was a macro simply declared
as "N" which caused a conflict somewhere within the cuda.h header file.

When the "N" macro was renamed to "BRK_N" (Brick N) and several source
modules were refactored, the error no longer occured.