[CMakeLists.txt] Add option for test install dir
The test programs built by the project should be installed in order to
test the success of the build. The BRICK_INSTALL_TESTDIR option has been
provided to specify the location of the test programs. Currently this
installs `single/cpu` in the test directory specified. The option is
used during the configuration stage:

    cmake -DBRICK_INSTALL_TESTDIR=<path-to-testdir>

During the install stage the test program will be installed to:

    <path-to-testdir>/single/cpu